+++ src/App.vue
@@ -1,4 +1,12 @@
-<script setup lang="ts"></script>
+<script setup lang="ts">
+// 重整頁面，需要重新登入
+// onMounted(() => {
+//   const userStore = useUserStore()
+//   if (!userStore.userInfo.userId) {
+//     router.replace(`/`)
+//   }
+// })
+</script>
 
 <template>
   <!-- <header>