+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.5.1] - 2025-08-01
+
+### Added
+
+- Implement rematch flow
+
 ## [0.5.0] - 2025-07-28
 
 ### Added